rename `test_gen/combine_observable_test` to `test_gen/combine_test`
