+++ test/test_gen/combine_test/test_cases/test_emit_if_all_children_emitted.dart
